refactor(config): split settings sections
